refactor: 함수명, 클래스명, 변수명 수정
