+++ app/src/main/java/woowacourse/shopping/ShoppingApplication.kt
@@ -1,17 +1,15 @@
 package woowacourse.shopping
 
 import android.app.Application
-import woowacourse.shopping.di.inject.CustomInject
+import woowacourse.shopping.di.inject.CustomInjector
 
 class ShoppingApplication : Application() {
 
+    lateinit var inject: CustomInjector
+        private set
+
     override fun onCreate() {
         super.onCreate()
-        inject = CustomInject()
-    }
-
-    companion object {
-        lateinit var inject: CustomInject
-            private set
+        inject = CustomInjector()
     }
 }
