+++ app/src/main/java/woowacourse/shopping/di/container/RepositoryContainer.kt
@@ -1,15 +1,15 @@
 package woowacourse.shopping.di.container
 
-import woowacourse.shopping.data.LocalCartRepository
-import woowacourse.shopping.data.LocalProductRepository
-import woowacourse.shopping.model.repository.CartRepository
-import woowacourse.shopping.model.repository.ProductRepository
+import woowacourse.shopping.data.DefaultCartRepository
+import woowacourse.shopping.data.DefaultProductRepository
+import woowacourse.shopping.repository.CartRepository
+import woowacourse.shopping.repository.ProductRepository
 
 object RepositoryContainer {
     private val productRepository: ProductRepository by lazy {
-        LocalProductRepository()
+        DefaultProductRepository()
     }
     private val cartRepository: CartRepository by lazy {
-        LocalCartRepository()
+        DefaultCartRepository()
     }
 }
